requirements.txt
streamlit  # for frontend 
google-generativeai # this help to access all the gemini pro APIs
python-dotenv    #load the environment variables    
